+++ datadog_checks_dev/datadog_checks/dev/tooling/commands/release/trello/testable.py
@@ -256,7 +256,7 @@ def testable(
     if repo == 'integrations-core':
         options = {
             '1': 'Integrations',
-            '2': 'Infra Integrations',
+            '2': 'Infra-Integrations',
             '3': 'Containers',
             '4': 'Core',
             '5': 'Platform',
@@ -274,7 +274,7 @@ def testable(
             '6': 'Processes',
             '7': 'Trace',
             '8': 'Integrations',
-            '9': 'Infra Integrations',
+            '9': 'Infra-Integrations',
             '10': 'Tools and Libraries',
             's': 'Skip',
             'q': 'Quit',