+++ pan_firewall/datadog_checks/pan.firewall/__init__.py
@@ -1,0 +1,6 @@
+# (C) Datadog, Inc. 2020-present
+# All rights reserved
+# Licensed under a 3-clause BSD style license (see LICENSE)
+from .__about__ import __version__
+
+__all__ = ['__version__']