+++ tls/datadog_checks/tls/tls.py
@@ -148,19 +148,23 @@ class TLSCheck(AgentCheck):
             try:
                 validator(cert, text_type(self._server_hostname))
             except service_identity.VerificationError:
+                message = 'The {} on the certificate does not match the given host'.format(host_type)
+                self.log.debug(message)
                 self.service_check(
                     SERVICE_CHECK_VALIDATION,
                     self.CRITICAL,
                     tags=self._tags,
-                    message='The {} on the certificate does not match the given host'.format(host_type),
+                    message=message,
                 )
                 return
             except service_identity.CertificateError as e:  # no cov
+                message = 'The certificate contains invalid/unexpected data: {}'.format(e)
+                self.log.debug(message)
                 self.service_check(
                     SERVICE_CHECK_VALIDATION,
                     self.CRITICAL,
                     tags=self._tags,
-                    message='The certificate contains invalid/unexpected data: {}'.format(e),
+                    message=message,
                 )
                 return
         self.log.debug('Certificate is valid')